Move DeviceSettingsManager to view package (addressing PR comments)
